authorization implemented

+++ src/api/notes/handler.js
@@ -16,8 +16,9 @@ class NotesHandler {
         try {
             this._validator.validateNotePayload(request.payload);
             const {title = 'untitled', body, tags} = request.payload;
+            const {id : credentialId} = request.auth.credentials;
 
-            const noteId = await this._services.addNote({title, body, tags});
+            const noteId = await this._services.addNote({title, body, tags, owner: credentialId});
     
             const response = h.response({
                 status: 'success',
@@ -49,7 +50,8 @@ class NotesHandler {
     }
 
     async getNotesHandler(request, h) {
-        const notes = await this._services.getNotes();
+        const { id } = request.auth.credentials;
+        const notes = await this._services.getNotes(id);
         const response = h.response({
             status: 'success',
             data: {
@@ -63,6 +65,8 @@ class NotesHandler {
     async getNoteByIdHandler(request, h) {
         try {
             const {id} = request.params;
+            const { id: credentialId } = request.auth.credentials;
+            await this._services.verifyNoteOwner(id, credentialId);
             const note = await this._services.getNoteById(id);
             return {
                 status: 'success',
@@ -94,6 +98,8 @@ class NotesHandler {
         try {
             this._validator.validateNotePayload(request.payload);
             const {id} = request.params;
+            const { id: credentialId } = request.auth.credentials;
+            await this._services.verifyNoteOwner(id, credentialId);
             await this._services.editNoteById(id, request.payload);
 
             return {
@@ -123,6 +129,8 @@ class NotesHandler {
     async deleteNoteByIdHandler(request, h) {
         try {
             const {id} = request.params;
+            const { id: credentialId } = request.auth.credentials;
+            await this._services.verifyNoteOwner(id, credentialId);
             await this._services.deleteNoteById(id);
 
             return {
